backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pymongo
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
cloudinary
razorpay
email-validator
google-generativeai
